+++ packages/cms-lib/oauth.js
@@ -1,6 +1,7 @@
 const OAuth2Manager = require('@hubspot/api-auth-lib/OAuth2Manager');
-const { updatePortalConfig } = require('./lib/config');
-const { logger } = require('./logger');
+const { updatePortalConfig, getPortalConfig } = require('./lib/config');
+const { logger, logErrorInstance } = require('./logger');
+const { AUTH_METHODS } = require('./lib/constants');
 
 const oauthManagers = new Map();
 
@@ -31,6 +32,34 @@ const getOauthManager = (portalId, portalConfig) => {
   return oauthManagers.get(portalId);
 };
 
+const setupOauth = (portalId, portalConfig) => {
+  const config = getPortalConfig(portalId) || {};
+  return new OAuth2Manager(
+    {
+      ...portalConfig,
+      environment: config.env || 'prod',
+    },
+    logger
+  );
+};
+
+const addOauthToPortalConfig = (portalId, oauth) => {
+  logger.log('Updating configuration');
+  try {
+    updatePortalConfig({
+      ...oauth.toObj(),
+      authType: AUTH_METHODS.oauth.value,
+      portalId,
+    });
+    logger.log('Configuration updated');
+    process.exit();
+  } catch (err) {
+    logErrorInstance(err);
+  }
+};
+
 module.exports = {
   getOauthManager,
+  setupOauth,
+  addOauthToPortalConfig,
 };
